chore(linkedin): share safety URL decoder
