+++ clis/linkedin/profile-projects.test.js
@@ -3,7 +3,7 @@ import { getRegistry } from '@jackwener/opencli/registry';
 import { CommandExecutionError } from '@jackwener/opencli/errors';
 import './profile-projects.js';
 
-const { normalizeProfileUrl, profileProjectsUrl, parseProjectText, parseProjectsSectionText, decodeLinkedInSafetyUrl, normalizeProject } = await import('./profile-projects.js').then((m) => m.__test__);
+const { normalizeProfileUrl, profileProjectsUrl, parseProjectText, parseProjectsSectionText, normalizeProject } = await import('./profile-projects.js').then((m) => m.__test__);
 
 describe('linkedin profile-projects adapter', () => {
   const command = getRegistry().get('linkedin/profile-projects');
@@ -102,10 +102,4 @@ Who your viewers also viewed`, 'https://www.linkedin.com/in/gauravsaxena1997/');
       .toMatchObject({ urls: 'https://example.com/demo' });
   });
 
-  it('decodes LinkedIn safety redirect URLs', () => {
-    expect(decodeLinkedInSafetyUrl('https://www.linkedin.com/safety/go/?url=https%3A%2F%2Fgithub.com%2Fjackwener%2FOpenCLI&urlhash=x'))
-      .toBe('https://github.com/jackwener/OpenCLI');
-    expect(decodeLinkedInSafetyUrl('https://www.linkedin.com/safety/go/?url=javascript%3Aalert(1)&urlhash=x'))
-      .toBe('');
-  });
 });
